Code review: 329770043: Moved EventBuffer into psort engine #771
