+++ config/dpkg/changelog
@@ -2,4 +2,4 @@ plaso (20170712-1) unstable; urgency=low
 
   * Auto-generated
 
- -- Log2Timeline <log2timeline-dev@googlegroups.com>  Wed, 12 Jul 2017 06:26:27 +0200
+ -- Log2Timeline <log2timeline-dev@googlegroups.com>  Wed, 12 Jul 2017 07:18:21 +0200
